dash_app/requirements.txt
cairosvg==2.5.2
dash==2.0.0
dash_bootstrap_components==1.0.2
ipykernel==6.7.0
lib==4.0.0
matplotlib==3.5.1
Pillow==9.0.0
plotly==5.5.0
protobuf==3.19.3
scikit_learn==1.0.2
scipy==1.7.3
statsmodels==0.13.1
torch==1.10.1
torchvision==0.11.2
tqdm==4.62.3